pyproject.toml
[project]
name = "datakura"
version = "0.0.1dev1"
description = "datakura"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.2.6",
    "django-ninja>=1.4.3",
    "loguru>=0.7.3",
    "pyjwt>=2.10.1",
    "python-decouple>=3.8",
]

[dependency-groups]
dev = [
    "django-stubs>=5.2.7",
    "factory-boy>=3.3.3",
    "mypy>=1.18.2",
    "pytest-django>=4.11.1",
    "ruff>=0.14.1",
]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "conf.settings.devel"
python_files = "tests.py test_*.py *_tests.py"


[tool.mypy]
python_version = "3.13"
plugins = ["mypy_django_plugin.main"]
files = "apps/**/*.py"
exclude = "(^conf/)"


[[tool.mypy.overrides]]
module = ["conf.*"]
ignore_errors = true

[[tool.mypy.overrides]]
module = ["decouple"]
ignore_missing_imports = true


[tool.django-stubs]
django_settings_module = "conf.settings.devel"
